[contour] Ensure font size (in config) can be specified as floating point value.

+++ src/contour/Config.h
@@ -91,7 +91,7 @@ struct TerminalProfile {
     int historyScrollMultiplier;
     bool autoScrollOnUpdate;
 
-    short fontSize;
+    double fontSize;
     FontSpecList fonts;
 
     int tabWidth;
